relay: fix tcpKeepAlive
